+++ components/merkledb/src/indexes/entry.rs
@@ -16,52 +16,44 @@
 
 use std::marker::PhantomData;
 
-use exonum_crypto::Hash;
-
 use crate::{
     access::{Access, AccessError, FromAccess},
-    views::{IndexAddress, IndexState, IndexType, RawAccess, RawAccessMut, View, ViewWithMetadata},
-    BinaryValue, ObjectHash,
+    views::{IndexAddress, IndexType, RawAccess, RawAccessMut, View, ViewWithMetadata},
+    BinaryValue,
 };
 
-/// A hashed index that may only contain one element.
+/// An index that may only contain one element.
 ///
 /// You can add an element to this index and check whether it exists. A value
-/// should implement [`BinaryValue`] and [`ObjectHash`] traits. Unlike [`Entry`],
-/// hashed entries are eagerly hashed and may participate in [state aggregation].
+/// should implement [`BinaryValue`] trait.
 ///
-/// [`BinaryValue`]: trait.BinaryValue.html
-/// [`ObjectHash`]: trait.ObjectHash.html
-/// [`Entry`]: struct.Entry.html
-/// [state aggregation]: index.html#state-aggregation
+/// [`BinaryValue`]: ../trait.BinaryValue.html
 #[derive(Debug)]
-pub struct ProofEntry<T: RawAccess, V> {
+pub struct Entry<T: RawAccess, V> {
     base: View<T>,
-    state: IndexState<T, Hash>,
     _v: PhantomData<V>,
 }
 
-impl<T, V> FromAccess<T> for ProofEntry<T::Base, V>
+impl<T, V> FromAccess<T> for Entry<T::Base, V>
 where
     T: Access,
-    V: BinaryValue + ObjectHash,
+    V: BinaryValue,
 {
     fn from_access(access: T, addr: IndexAddress) -> Result<Self, AccessError> {
-        let view = access.get_or_create_view(addr, IndexType::ProofEntry)?;
+        let view = access.get_or_create_view(addr, IndexType::Entry)?;
         Ok(Self::new(view))
     }
 }
 
-impl<T, V> ProofEntry<T, V>
+impl<T, V> Entry<T, V>
 where
     T: RawAccess,
     V: BinaryValue,
 {
-    pub(crate) fn new(view: ViewWithMetadata<T>) -> Self {
-        let (base, state) = view.into_parts();
+    fn new(view: ViewWithMetadata<T>) -> Self {
+        let base = view.into();
         Self {
             base,
-            state,
             _v: PhantomData,
         }
     }
@@ -75,7 +67,7 @@ where
     ///
     /// let db = TemporaryDB::new();
     /// let fork = db.fork();
-    /// let mut index = fork.get_proof_entry("name");
+    /// let mut index = fork.get_entry("name");
     /// assert_eq!(None, index.get());
     ///
     /// index.set(10);
@@ -94,7 +86,7 @@ where
     ///
     /// let db = TemporaryDB::new();
     /// let fork = db.fork();
-    /// let mut index = fork.get_proof_entry("name");
+    /// let mut index = fork.get_entry("name");
     /// assert!(!index.exists());
     ///
     /// index.set(10);
@@ -105,10 +97,10 @@ where
     }
 }
 
-impl<T, V> ProofEntry<T, V>
+impl<T, V> Entry<T, V>
 where
     T: RawAccessMut,
-    V: BinaryValue + ObjectHash,
+    V: BinaryValue,
 {
     /// Changes a value of the entry.
     ///
@@ -119,14 +111,13 @@ where
     ///
     /// let db = TemporaryDB::new();
     /// let fork = db.fork();
-    /// let mut index = fork.get_proof_entry("name");
+    /// let mut index = fork.get_entry("name");
     ///
     /// index.set(10);
     /// assert_eq!(Some(10), index.get());
     /// ```
     pub fn set(&mut self, value: V) {
-        self.state.set(value.object_hash());
-        self.base.put(&(), value);
+        self.base.put(&(), value)
     }
 
     /// Removes a value of the entry.
@@ -138,7 +129,7 @@ where
     ///
     /// let db = TemporaryDB::new();
     /// let fork = db.fork();
-    /// let mut index = fork.get_proof_entry("name");
+    /// let mut index = fork.get_entry("name");
     ///
     /// index.set(10);
     /// assert_eq!(Some(10), index.get());
@@ -147,8 +138,7 @@ where
     /// assert_eq!(None, index.get());
     /// ```
     pub fn remove(&mut self) {
-        self.state.unset();
-        self.base.remove(&());
+        self.base.remove(&())
     }
 
     /// Takes the value out of the entry, leaving a None in its place.
@@ -160,7 +150,7 @@ where
     ///
     /// let db = TemporaryDB::new();
     /// let fork = db.fork();
-    /// let mut index = fork.get_proof_entry("name");
+    /// let mut index = fork.get_entry("name");
     ///
     /// index.set(10);
     /// assert_eq!(Some(10), index.get());
@@ -186,7 +176,7 @@ where
     ///
     /// let db = TemporaryDB::new();
     /// let fork = db.fork();
-    /// let mut index = fork.get_proof_entry("name");
+    /// let mut index = fork.get_entry("name");
     ///
     /// index.set(10);
     /// assert_eq!(Some(10), index.get());
@@ -199,98 +189,5 @@ where
         let previous = self.get();
         self.set(value);
         previous
-    }
-}
-
-/// `object_hash` is computed as SHA-256 of the entry serialization, or `Hash::zero()` if
-/// the entry is not set.
-///
-/// # Examples
-///
-/// ```
-/// # use exonum_merkledb::{access::AccessExt, TemporaryDB, Database, Entry, ObjectHash};
-/// # use exonum_crypto::{self, Hash};
-/// let db = TemporaryDB::new();
-/// let fork = db.fork();
-/// let mut index = fork.get_proof_entry("name");
-/// assert_eq!(Hash::default(), index.object_hash());
-///
-/// let value = 10;
-/// index.set(value);
-/// assert_eq!(exonum_crypto::hash(&[value]), index.object_hash());
-/// ```
-impl<T, V> ObjectHash for ProofEntry<T, V>
-where
-    T: RawAccess,
-    V: BinaryValue + ObjectHash,
-{
-    fn object_hash(&self) -> Hash {
-        self.state.get().unwrap_or_default()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{access::AccessExt, Database, TemporaryDB};
-    use std::borrow::Cow;
-
-    #[test]
-    fn basics() {
-        let db = TemporaryDB::new();
-        let fork = db.fork();
-        {
-            let mut entry = fork.get_proof_entry("test");
-            assert!(!entry.exists());
-            entry.set(25_u64);
-            assert!(entry.exists());
-            assert_eq!(entry.get(), Some(25));
-            assert_eq!(entry.swap(42), Some(25));
-        }
-        db.merge(fork.into_patch()).unwrap();
-
-        let snapshot = db.snapshot();
-        let entry = snapshot.get_proof_entry::<_, u64>("test");
-        assert_eq!(entry.get(), Some(42));
-        assert_eq!(entry.object_hash(), 42_u64.object_hash());
-    }
-
-    #[test]
-    fn entry_with_custom_hashing() {
-        #[derive(Debug, PartialEq)]
-        struct CustomHash(u8);
-
-        impl BinaryValue for CustomHash {
-            fn to_bytes(&self) -> Vec<u8> {
-                vec![self.0]
-            }
-
-            fn from_bytes(bytes: Cow<'_, [u8]>) -> Result<Self, failure::Error> {
-                u8::from_bytes(bytes).map(CustomHash)
-            }
-        }
-
-        impl ObjectHash for CustomHash {
-            fn object_hash(&self) -> Hash {
-                Hash::new([self.0; exonum_crypto::HASH_SIZE])
-            }
-        }
-
-        let db = TemporaryDB::new();
-        let fork = db.fork();
-        {
-            let mut entry = fork.get_proof_entry("test");
-            entry.set(CustomHash(11));
-            assert!(entry.exists());
-            assert_eq!(entry.get(), Some(CustomHash(11)));
-        }
-        db.merge(fork.into_patch()).unwrap();
-
-        let snapshot = db.snapshot();
-        let entry = snapshot.get_proof_entry::<_, ()>("test");
-        assert_eq!(
-            entry.object_hash(),
-            Hash::new([11; exonum_crypto::HASH_SIZE])
-        );
     }
 }